Fix the issue of lines between tile boundaries.
The cause of these lines is the background showing between tile boundaries. The current background color if gray, so the lines appear as a shade of gray. If the background color is changed to red, the lines appear as shades of red.

I'm not exactly sure why these lines appear. It has something to do with how texture coordinates are clamped to the border. The OpenGL ES specs that WebGL is derived from state the behavior that texture clamping should have (Section 3.7.6 in the OpenGL ES 2.0 spec and section 8.13.2 in the OpenGL ES spec; the devices for implementing this clamping differ between these two versions but should be equivalent).

Although WebGL is defined as a JavaScript API of OpenGL ES, the implementations we run on are generally on the desktop, which provides full OpenGL (generally version 4 or thereabouts) support but not OpenGL ES support. Therefore, the WebGL implementations need to be translated into full OpenGL and errors may have been introduced.

It has been noted that this bug is also visible on Android, where the browser should be built on top of a valid OpenGL ES implementation. It has also been noted that the bug does not manifest itself on iOS, which also has a true OpenGL ES implmentation. I don't know what to make of this data.

The solution I implemented was to make texel coordinates so that they align at texel centers at the ends. Thus, for a texture map with N pixels on each edge (where it is assumed that N is a power of two, although my change does NOT reply on that assumption), one can map the extreme texture coordinates to the center of the extreme texels. Thus
 u(or v) = 0 => texel sample at 0.5
 u(or v) = 1 => texel sample at N - 0.5
 To get the uv coordinates to sample these extreme texel locations, u and v are scaled by (N-1)/N and offset by 1/2N.

git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@2647 822214bb-dcf7-4c8d-a767-17ae0fb20d88