Create a Renderable called UserFacingText and a Layer called PlacenameLayer, as well as a toy demo call Peaks.
UserFacingText uses SVG text to create a texture map containing the text associated with the UserFacingText object. This texture map is then drawn into a rectangle centered on the screen location of the position specified during creation of the UserFacingText.

Also included is a little demo that displays UserFacingText in a place name layer for the most "ultra-prominent" peak in each state for those states in the US that have them.

Known issues:
   1) Trying to draw a black boundary around white text isn;t being drawn.
   2) Far clipping isn't effective enough. Text will go over the horizon and eventually disappear, but this happens well after the position is no longer visible.
   
Validated to work correctly on the following browsers:
   OS X:
      Safari
      Firefox
      Chrome
   Win7:
      IE 11
      Firefox
      Chrome
   iOS:
      Safari
      Chrome
      Coast


git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@2658 822214bb-dcf7-4c8d-a767-17ae0fb20d88